2003-03-24: change in the semantics of empty parameter lists
Previously invoking a macro with an empty parameter list <A()> (meta charact
was different from invoking it without a parameterlist <A>; in the first cas
retained the the values of the last invocation in the current chunk of a mac
In the second case all the parameters hava a null value. Both cases have bee
made equivalent.