add compiler option
To use non ASCII characters for the meta-character in some systems the char
type must be unsigned. That allows the use of eight bit encodings (such as ISO-8859-1)
for the metacharacter.

Locale settings (LC_ALL or LC_CTYPE) may also be necessary at runtime to use non-ASCII
characters as scrap names, etc.